Enhance ProjectSettingsMenu and add visibility converter
Updated ProjectSettingsMenu.xaml to include a new resource dictionary and a custom ListView style. Modified layout to add a GroupBox for the current project and updated the ItemTemplate to display the UpdatedAt property with an icon.

Removed unused folders from DPUnity.Wpf.Controls.csproj and added a new BooleanToHiddenVisibilityConverter class to handle boolean to visibility conversions.