Simplify window owner condition and update version
Modified `DPDialog.cs` to allow setting the window owner if it is loaded, regardless of visibility. This enhances the dialog service's behavior.

Updated version in `DPUnity.Wpf.Controls.csproj` from `1.0.16-preview` to `1.0.16`, marking a stable release.

Refactor dialog owner logic and update version number

Simplified the condition for setting the window owner in `DPDialog.cs` by removing the visibility check, allowing for improved usability of the dialog service. Adjusted logic for finding an active window to enhance functionality. Updated the version in `DPUnity.Wpf.Controls.csproj` from `1.0.16-preview` to `1.0.16`, marking the transition to a stable release.